Starting with code stolen from my process-calendar project (also on Google code)
Did a basic test to make sure that I can access the RSS feed of chess.com.  

Login Form has been created. I haven't gone over the rest of the objects imported over from the process-calendar app, so things are a bit of a mess at present.

App does not work when you run it yet, so don't try.